Complete overhaul of process tracking system Instead of holding a process state for each process event, all process info is held in one structure. Process info that changes throughout time is represented in a binary tree indexed by the start time of that piece of info. Process info dissection displays all process info across its lifespan, while annotating relevance of time-dependent pieces of info, and referencing events in the process' lifecycle.
In order to avoid messy and innacurate logic regarding the order in
which process events are encountered (and how to decide when info
becomes valid/invalid), a requirement was added for
process tracking - the capture must be ordered chronologically.
A capture can be ordered using the `reordercap` utility.

Right now, only fork events contribute to process tracking.
Other events will be added next.

+++ wiretap/trace-cmd.c
@@ -1156,7 +1156,7 @@ static inline enum get_record_result __tracecmd_get_record(FILE_T fh, struct tra
             if (!wtap_read_bytes(fh, &buf, sizeof(guint32), err, err_info))
                 return GET_RECORD_RESULT_ERROR;
             time_delta_high = tracecmd->big_endian ? pntoh32(buf) : pletoh32(buf);
-            state->current_ts += (time_delta_high << 27) + header.time_delta;
+            state->current_ts += ((guint64)time_delta_high << 27) + header.time_delta;
 
             return GET_RECORD_RESULT_NOT_EVENT;
